+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.16] - 2025-05-21
+
+### Changed
+
+- Update module to version 0.2.16
+- Enhance localization strings
+- Improve output messages in New-ADDiagram function
+
 ## [0.2.15] - 2025-05-10
 
 ### Changed